Move the admin settings to the workflow section
