--- settings/admin.php
@@ -1,27 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2016 Joas Schilling <coding@schilljs.com>
- *
- * @license GNU AGPL version 3 or any later version
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-$app = new \OCA\FilesAutomatedTagging\AppInfo\Application();
-$container = $app->getContainer();
-/** @var \OCP\AppFramework\Http\TemplateResponse $response */
-$response = $container->query('OCA\FilesAutomatedTagging\Controller\AdminController')->index();
-
-return $response->render();
